# src/docs/openapi/openapi.yaml
openapi: 3.0.0
info:
  title: Block Explorer API
  version: 1.0.0
  description: API documentation for the Block Explorer Interface

servers:
  - url: /api
    description: Default API endpoint

components:
  schemas:
    Block:
      type: object
      properties:
        id:
          type: string
          example: "0x123..."
        height:
          type: integer
          example: 12345
        datetime_added:
          type: string
          format: date-time
        hash:
          type: string
          example: "0xabc..."
        miner:
          type: string
          example: "0xdef..."
        tx_count:
          type: integer
          example: 150

    Algorithm:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        datetime_added:
          type: string
          format: date-time
        player_id:
          type: string
        challenge_id:
          type: string
        tx_hash:
          type: string

    Benchmark:
      type: object
      properties:
        id:
          type: string
        datetime_added:
          type: string
          format: date-time
        merkle_root:
          type: string
          nullable: true
        num_solutions:
          type: integer
        block_confirmed:
          type: integer

    PaginatedResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Block'
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer

    Error:
      type: object
      properties:
        message:
          type: string
        code:
          type: string
        details:
          type: object

  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 0
      description: Page number for pagination
    
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 10
      description: Number of items per page

paths:
  /blocks:
    get:
      summary: Get blocks list
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /blocks/{id}:
    get:
      summary: Get block details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Block'
        '404':
          description: Block not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /algorithms:
    get:
      summary: Get algorithms list
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'

  /benchmarks:
    get:
      summary: Get benchmarks list
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'